// src/types/openapi.d.ts
/**
 * This file was auto-generated by openapi-typescript.
 * Do not make direct changes to the file.
 */


export interface paths {
  "/user": {
    /** Get all users */
    get: {
      responses: {
        /** @description successful operation */
        200: {
          content: {
            "application/json": unknown;
          };
        };
        /** @description Document not found */
        400: {
          content: never;
        };
      };
    };
  };
  "/user/{userId}": {
    /** Get user detail */
    get: {
      parameters: {
        query: {
          /** @description User id for define identifier */
          userId: number;
        };
      };
      responses: {
        /** @description successful operation */
        200: {
          content: {
            "application/json": unknown;
          };
        };
        /** @description Invalid user id */
        400: {
          content: never;
        };
      };
    };
  };
}

export type webhooks = Record<string, never>;

export interface components {
  schemas: {
    Users: {
      userId?: number;
      username: string;
      age: number;
      /** Format: email */
      email: string;
      point?: number;
      /** @enum {string} */
      status: "active" | "notActive";
      /** @enum {string} */
      gender: "male" | "female";
      savings?: number;
    };
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}

export type $defs = Record<string, never>;

export type external = Record<string, never>;

export type operations = Record<string, never>;